Feat : Detail의 Additional, Overview 분리
