+++ server/view/index.html
@@ -7,5 +7,5 @@
 </head>
 <body>
     <div id='root'></div>
-<script src="react.bundle.6044c4f2023d4d6b5af8.js"></script><script src="vendors.6044c4f2023d4d6b5af8.js"></script><script src="app.6044c4f2023d4d6b5af8.js"></script></body>
+<script src="react.bundle.e8a64c57849fe7d26872.js"></script><script src="vendors.e8a64c57849fe7d26872.js"></script><script src="app.e8a64c57849fe7d26872.js"></script></body>
 </html>
